Move reservoir sampling into its own file
