+++ backend/app.py
@@ -6,6 +6,7 @@ import csv
 from tempfile import NamedTemporaryFile
 from urllib.parse import parse_qs
 from chalicelib import pdftotxt
+from chalicelib.algo import reservior_sampling
 from chalicelib.train import export_model
 import cgi
 import boto3
@@ -13,7 +14,6 @@ import logging
 from sklearn.externals import joblib
 from sklearn import metrics
 import pandas as pd
-import random
 from dateparser.search import search_dates
 from datetime import datetime
 import hashlib
@@ -106,26 +106,6 @@ def dynamodb_response_to_df(response):
                 return value
         df = df.applymap(convert_dynamo_data_type)
     return df
-
-
-# new_data and existing_samples should be of type pd.DataFrame
-def reservior_sampling(sample_size, new_data,
-                       existing_record_count=0, existing_samples=pd.DataFrame()):
-    new_data = new_data.reset_index(drop=True)
-    samples = existing_samples.reset_index(drop=True)
-    replaced_indexes = {}
-    for index, record in new_data.iterrows():
-        if samples.shape[0] < sample_size:
-            samples = samples.append(record, ignore_index=True)
-            replaced_indexes[index] = index
-        else:
-            r = random.randint(0, existing_record_count + index)
-            if r < sample_size:
-                samples.loc[r] = record
-                replaced_indexes[r] = index
-    samples.reset_index(inplace=True)
-    remaining_indexes = set(range(len(new_data))) - set(replaced_indexes.values())
-    return (samples, remaining_indexes)
 
 
 def dataframe_as_response(df, accept_header):
